# reduce cost of pipeline stages by pooling maps we create for labels