--- a/pkg/logql/log/base
+++ b/pkg/logql/log/base
@@ -1,0 +1,12 @@
+goos: linux
+goarch: amd64
+pkg: github.com/grafana/loki/pkg/logql/log
+cpu: AMD Ryzen 9 5950X 16-Core Processor            
+Benchmark_Pipeline/pipeline_bytes-32                                        	  153394	      7749 ns/op	    5082 B/op	      37 allocs/op
+Benchmark_Pipeline/pipeline_string-32                                       	  153319	      7691 ns/op	    5146 B/op	      38 allocs/op
+Benchmark_Pipeline/line_extractor_bytes-32                                  	  148069	      7922 ns/op	    5151 B/op	      37 allocs/op
+Benchmark_Pipeline/line_extractor_string-32                                 	  149070	      7900 ns/op	    5149 B/op	      37 allocs/op
+Benchmark_Pipeline/label_extractor_bytes-32                                 	  146860	      7952 ns/op	    5148 B/op	      37 allocs/op
+Benchmark_Pipeline/label_extractor_string-32                                	  147686	      7940 ns/op	    5148 B/op	      37 allocs/op
+PASS
+ok  	github.com/grafana/loki/pkg/logql/log	190.270s