--- a/pkg/logql/log/fmt.go
+++ b/pkg/logql/log/fmt.go
@@ -221,7 +221,13 @@ func (lf *LineFormatter) Process(ts int64, line []byte, lbs *LabelsBuilder) ([]b
 	lf.currentLine = line
 	lf.currentTs = ts
 
-	if err := lf.Template.Execute(lf.buf, lbs.Map()); err != nil {
+	// map now is taking from a pool
+	m := lbs.Map()
+	defer func() {
+		clear(m)
+		stringMapPool.Put(m)
+	}()
+	if err := lf.Template.Execute(lf.buf, m); err != nil {
 		lbs.SetErr(errTemplateFormat)
 		lbs.SetErrorDetails(err.Error())
 		return line, true
@@ -380,7 +386,7 @@ func (lf *LabelsFormatter) Process(ts int64, l []byte, lbs *LabelsBuilder) ([]by
 	lf.currentLine = l
 	lf.currentTs = ts
 
-	var data interface{}
+	var data map[string]string
 	for _, f := range lf.formats {
 		if f.Rename {
 			v, category, ok := lbs.GetWithCategory(f.Value)
@@ -393,6 +399,10 @@ func (lf *LabelsFormatter) Process(ts int64, l []byte, lbs *LabelsBuilder) ([]by
 		lf.buf.Reset()
 		if data == nil {
 			data = lbs.Map()
+			defer func() {
+				clear(data)
+				stringMapPool.Put(data)
+			}()
 		}
 		if err := f.tmpl.Execute(lf.buf, data); err != nil {
 			lbs.SetErr(errTemplateFormat)