--- a/pkg/logql/log/fmt.go
+++ b/pkg/logql/log/fmt.go
@@ -224,7 +224,6 @@ func (lf *LineFormatter) Process(ts int64, line []byte, lbs *LabelsBuilder) ([]b
 	// map now is taking from a pool
 	m := lbs.Map()
 	defer func() {
-		clear(m)
 		stringMapPool.Put(m)
 	}()
 	if err := lf.Template.Execute(lf.buf, m); err != nil {
@@ -386,7 +385,12 @@ func (lf *LabelsFormatter) Process(ts int64, l []byte, lbs *LabelsBuilder) ([]by
 	lf.currentLine = l
 	lf.currentTs = ts
 
-	var data map[string]string
+	var data = stringMapPool.Get().(map[string]string)
+	clear(data)
+	
+	defer func() {
+		stringMapPool.Put(data)
+	}()
 	for _, f := range lf.formats {
 		if f.Rename {
 			v, category, ok := lbs.GetWithCategory(f.Value)
@@ -397,12 +401,8 @@ func (lf *LabelsFormatter) Process(ts int64, l []byte, lbs *LabelsBuilder) ([]by
 			continue
 		}
 		lf.buf.Reset()
-		if data == nil {
-			data = lbs.Map()
-			defer func() {
-				clear(data)
-				stringMapPool.Put(data)
-			}()
+		if len(data) == 0 {
+			lbs.IntoMap(data)
 		}
 		if err := f.tmpl.Execute(lf.buf, data); err != nil {
 			lbs.SetErr(errTemplateFormat)