--- a/pkg/logql/log/labels.go
+++ b/pkg/logql/log/labels.go
@@ -442,8 +442,8 @@ type stringMapPool struct {
 	pool sync.Pool
 }
 
-func newStringMapPool() stringMapPool {
-	return stringMapPool{
+func newStringMapPool() *stringMapPool {
+	return &stringMapPool{
 		pool: sync.Pool{
 			New: func() interface{} {
 				return make(map[string]string)
@@ -452,13 +452,13 @@ func newStringMapPool() stringMapPool {
 	}
 }
 
-func (s stringMapPool) Get() map[string]string {
+func (s *stringMapPool) Get() map[string]string {
 	m := s.pool.Get().(map[string]string)
 	clear(m)
 	return m
 }
 
-func (s stringMapPool) Put(m map[string]string) {
+func (s *stringMapPool) Put(m map[string]string) {
 	s.pool.Put(m)
 }