--- a/pkg/logql/log/last
+++ b/pkg/logql/log/last
@@ -1,0 +1,12 @@
+goos: linux
+goarch: amd64
+pkg: github.com/grafana/loki/pkg/logql/log
+cpu: AMD Ryzen 9 5950X 16-Core Processor            
+Benchmark_Pipeline/pipeline_bytes-32         	  172089	      6673 ns/op	    1447 B/op	      35 allocs/op
+Benchmark_Pipeline/pipeline_string-32        	  179335	      6670 ns/op	    1511 B/op	      36 allocs/op
+Benchmark_Pipeline/line_extractor_bytes-32   	  172724	      6853 ns/op	    1513 B/op	      35 allocs/op
+Benchmark_Pipeline/line_extractor_string-32  	  170238	      6866 ns/op	    1513 B/op	      35 allocs/op
+Benchmark_Pipeline/label_extractor_bytes-32  	  169942	      7004 ns/op	    1513 B/op	      35 allocs/op
+Benchmark_Pipeline/label_extractor_string-32 	  168538	      6982 ns/op	    1513 B/op	      35 allocs/op
+PASS
+ok  	github.com/grafana/loki/pkg/logql/log	7.650s