--- a/pkg/logql/log/mapPool-3
+++ b/pkg/logql/log/mapPool-3
@@ -1,0 +1,12 @@
+goos: linux
+goarch: amd64
+pkg: github.com/grafana/loki/pkg/logql/log
+cpu: AMD Ryzen 9 5950X 16-Core Processor            
+Benchmark_Pipeline/pipeline_bytes-32         	  180068	      6571 ns/op	    1455 B/op	      34 allocs/op
+Benchmark_Pipeline/pipeline_string-32        	  179401	      6597 ns/op	    1518 B/op	      35 allocs/op
+Benchmark_Pipeline/line_extractor_bytes-32   	  158112	      6923 ns/op	    1521 B/op	      34 allocs/op
+Benchmark_Pipeline/line_extractor_string-32  	  150291	      7061 ns/op	    1521 B/op	      34 allocs/op
+Benchmark_Pipeline/label_extractor_bytes-32  	  166911	      7037 ns/op	    1521 B/op	      34 allocs/op
+Benchmark_Pipeline/label_extractor_string-32 	  169440	      6983 ns/op	    1522 B/op	      34 allocs/op
+PASS
+ok  	github.com/grafana/loki/pkg/logql/log	7.548s