@dannykopping 2023-12-15
Can you share a couple profiles showing this? (_probably best to export the pprof files from pyroscope and attach here instead of linking, for the sake of the community_)