@MasslessParticle 2023-12-15
This seems like a cool idea, in general. I've left a comment with a potential QOL improvement for someone using it.