@cstyan 2023-12-15
> Can you share a couple profiles showing this? (probably best to export the pprof files from pyroscope and attach here instead of linking, for the sake of the community)

I actually don't know how to export profiles from pyroscope still, but here's a screenshot.

![2023-12-15-142509_1201x300_scrot](https://github.com/grafana/loki/assets/3246492/bef7216b-6c7c-4a0c-b3e0-5b10dc30701b)